requirements.txt
pytest==6.2.2
requests==2.25.1
requests_mock
python-dateutil
pytest-html==3.1.1